requirements.txt
# Python 3.10+ recommended
Bootstrap-Flask==2.2.0
Flask==2.3.2
WTForms==3.0.1
Flask-WTF==1.2.1
Werkzeug==3.0.0
python-dotenv>=1.0.0   # loads SECRET_KEY from .env
